style: move SidebarSettings
