+++ wondrous-app/components/Common/Sidebar/index.tsx
@@ -1,4 +1,3 @@
 export { default as SideBarComponent } from './Main';
 export { default as EntitySidebar } from './Entity';
-export { default as SettingsWrapper } from './Settings';
 export { default as ResourcesSidebar } from './Resources';
