+++ wondrous-app/components/Settings/Members/index.tsx
@@ -17,7 +17,7 @@ import { SnackbarAlertContext } from 'components/Common/SnackbarAlert';
 import { TaskMenuIcon } from 'components/Icons/taskMenu';
 import ConfirmModal, { SubmitButtonStyle } from 'components/Common/ConfirmModal';
 import { NewOrgInviteLinkModal } from 'components/Common/NewInviteLinkModal/OrgInviteLink';
-import { SettingsWrapper } from 'components/Common/Sidebar';
+import SettingsWrapper from 'components/Common/SidebarSettings';
 import MemberRoles from '../MemberRoles';
 import MemberRoleDropdown from './MemberRoleDropdown';
 import InviteMember from './InviteMember';
